chore: update files after the rename
